% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdwtrsmpkwale-package.R
\docType{package}
\name{grdwtrsmpkwale-package}
\alias{grdwtrsmpkwale}
\alias{grdwtrsmpkwale-package}
\title{grdwtrsmpkwale: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openwashdata/grdwtrsmpkwale}
  \item Report bugs at \url{https://github.com/openwashdata/grdwtrsmpkwale/issues}
}

}
\author{
\strong{Maintainer}: Mian Zhong \email{mzhong@ethz.ch} (\href{https://orcid.org/0009-0009-4546-7214}{ORCID})

Authors:
\itemize{
  \item Sebastian Camilo Loos \email{sloos@ethz.ch} (\href{https://orcid.org/0000-0002-8830-1734}{ORCID})
}

}
\keyword{internal}
